Cache app db config to speed up checking account alias
AB_Banking_GetAccountSpecByAlias always reads the whole app config
causing slow operations when changing the ledger in kmymoney 4.8

AB_Banking_GetAccountSpecByAlias reads the configuration file for the
application in question on every call, which resulted in a very long
processing time (detected on an older laptop probably caused by file
locks).

To work around this problem, the file is only loaded on the first call
and stored in the AB_Banking instance for further access.

When the configuration file is updated, the read cache is cleared.